+++ Tests/FO-DICOM.Tests/IO/Reader/DicomDatasetReaderObserverTest.cs
@@ -12,7 +12,7 @@ using Xunit;
 namespace FellowOakDicom.Tests.IO.Reader
 {
 
-    [Collection("General")]
+    [Collection(TestCollections.General)]
     public class DicomDatasetReaderObserverTest
     {
         #region Unit tests