+++ Tests/FO-DICOM.Tests/Imaging/Mathematics/Vector3DTest.cs
@@ -9,7 +9,7 @@ using Xunit;
 namespace FellowOakDicom.Tests.Imaging.Mathematics
 {
 
-    [Collection("General")]
+    [Collection(TestCollections.General)]
     public class Vector3DTests
     {