+++ Tests/FO-DICOM.Tests/Media/DicomFileScannerTest.cs
@@ -8,7 +8,7 @@ using Xunit;
 namespace FellowOakDicom.Tests.Media
 {
 
-    [Collection("General")]
+    [Collection(TestCollections.General)]
     public class DicomFileScannerTest
     {
         [Fact]