+++ Tests/FO-DICOM.Tests/Network/DicomCFindRequestTest.cs
@@ -14,7 +14,7 @@ using Xunit;
 namespace FellowOakDicom.Tests.Network
 {
 
-    [Collection("Network")]
+    [Collection(TestCollections.Network)]
     public class DicomCFindRequestTest
     {
         #region Unit Tests